Implement everything with generics

+++ examples/total_stopping_time.rs
@@ -8,13 +8,13 @@ fn main() {
     let mut s: Vec<u128> = (1..=end).collect();
     let now = Instant::now();
     s.iter_mut()
-        .for_each(|p| *p = collatz::total_stopping_time(*p as u128));
+        .for_each(|p| *p = collatz::total_stopping_time(*p as u128) as u128);
     println!("Time single threaded: {}ms", now.elapsed().as_millis());
 
     let mut r: Vec<u128> = (1..=end).collect();
     let now = Instant::now();
     r.par_iter_mut()
-        .for_each(|p| *p = collatz::total_stopping_time(*p as u128));
+        .for_each(|p| *p = collatz::total_stopping_time(*p as u128) as u128);
     println!("Time with rayon: {}ms", now.elapsed().as_millis());
 
     // Verify the results are the same
